+++ packages/itwin/tree-widget/src/components/trees/category-tree/CategoriesTree.tsx
@@ -141,6 +141,7 @@ async function setViewType(activeView?: Viewport) {
     return;
 
   const view = activeView.view as SpatialViewState;
-  const viewType = view.is3d() ? "3d" : "2d";
+  // TODO: remove this eslint rule when tree-widget uses itwinjs-core 4.0.0 version
+  const viewType = view.is3d() ? "3d" : "2d"; // eslint-disable-line @itwin/no-internal
   await Presentation.presentation.vars(RULESET_CATEGORIES.id).setString("ViewType", viewType);
 }