+++ packages/itwin/tree-widget/src/test/trees/CategoriesVisibilityUtils.test.ts
@@ -5,45 +5,34 @@
 
 import * as sinon from "sinon";
 import * as moq from "typemoq";
+import { ECSqlReader, SubCategoryAppearance } from "@itwin/core-common";
 import {
-  IModelApp, IModelConnection, NoRenderApp, PerModelCategoryVisibility, ScreenViewport, SpatialViewState, SubCategoriesCache, ViewManager, Viewport,
-  ViewState,
+  IModelApp, IModelConnection, NoRenderApp, PerModelCategoryVisibility, ScreenViewport, SpatialViewState, ViewManager, Viewport, ViewState,
 } from "@itwin/core-frontend";
-import { KeySet } from "@itwin/presentation-common";
-import { Presentation, PresentationManager, SelectionChangeEvent, SelectionManager } from "@itwin/presentation-frontend";
 import { enableCategory, enableSubCategory, toggleAllCategories } from "../../components/trees/CategoriesVisibilityUtils";
-import { CategoryInfo } from "../../components/trees/category-tree/CategoryVisibilityHandler";
-import { mockPresentationManager, TestUtils } from "../TestUtils";
-import { ECSqlReader, QueryRowProxy } from "@itwin/core-common";
+import { TestUtils } from "../TestUtils";
 
 describe("CategoryVisibilityUtils", () => {
   before(async () => {
-    await NoRenderApp.startup();
+    // TODO: remove this eslint rule when tree-widget uses itwinjs-core 4.0.0 version
+    await NoRenderApp.startup(); // eslint-disable-line @itwin/no-internal
     await TestUtils.initialize();
   });
 
   after(async () => {
     TestUtils.terminate();
-    Presentation.terminate();
     await IModelApp.shutdown();
   });
 
-  afterEach(() => {
-    sinon.restore();
-  });
-
   const imodelMock = moq.Mock.ofType<IModelConnection>();
-  const selectionManagerMock = moq.Mock.ofType<SelectionManager>();
-  let presentationManagerMock: moq.IMock<PresentationManager>;
   const viewportMock = moq.Mock.ofType<Viewport>();
   const viewStateMock = moq.Mock.ofType<SpatialViewState>();
   const viewManagerMock = moq.Mock.ofType<ViewManager>();
   const selectedViewMock = moq.Mock.ofType<ScreenViewport>();
   const selectedViewStateMock = moq.Mock.ofType<ViewState>();
-  const perModelCategoryVisibilityMock = moq.Mock.ofType<PerModelCategoryVisibility.Overrides>();
-  const subCategoriesCacheMock = moq.Mock.ofType<SubCategoriesCache>();
   const queryReaderMock = moq.Mock.ofType<ECSqlReader>();
-  const queryRowProxyMock = moq.Mock.ofType<QueryRowProxy>();
+  const categoriesMock = moq.Mock.ofType<IModelConnection.Categories>();
+  const perModelCategoryVisibilityMock = moq.Mock.ofType<PerModelCategoryVisibility.Overrides>();
 
   const mockViewManagerForEachViewport = (viewport: Viewport, times = moq.Times.once()) => {
     viewManagerMock.reset();
@@ -54,57 +43,55 @@ describe("CategoryVisibilityUtils", () => {
       .verifiable(times);
   };
 
-  const categories: CategoryInfo[] = [
-    {
-      categoryId: "CategoryId",
-      subCategoryIds: ["SubCategoryId"],
-    },
-  ];
+  const categoryId = "CategoryId";
+  const subCategoryId = "SubCategoryId";
+  const categoriesInfo = new Map(
+    [[
+      categoryId,
+      {
+        id: categoryId,
+        subCategories: new Map(
+          [[
+            subCategoryId,
+            {
+              id: subCategoryId,
+              categoryId,
+              appearance: new SubCategoryAppearance(),
+            },
+          ]]
+        ),
+      },
+    ]]
+  );
 
   beforeEach(() => {
-    viewManagerMock.reset();
-    imodelMock.reset();
-    selectionManagerMock.reset();
-    viewportMock.reset();
-    viewStateMock.reset();
-    selectedViewMock.reset();
-    perModelCategoryVisibilityMock.reset();
-
-    const selectionChangeEvent = new SelectionChangeEvent();
-    selectionManagerMock.setup((x) => x.selectionChange).returns(() => selectionChangeEvent);
-    selectionManagerMock.setup((x) => x.getSelectionLevels(imodelMock.object)).returns(() => []);
-    selectionManagerMock.setup((x) => x.getSelection(imodelMock.object, moq.It.isAny())).returns(() => new KeySet());
-    Presentation.setSelectionManager(selectionManagerMock.object);
-
-    const mocks = mockPresentationManager();
-    presentationManagerMock = mocks.presentationManager;
-    Presentation.setPresentationManager(presentationManagerMock.object);
-
     imodelMock.setup((x) => x.createQueryReader(moq.It.isAny(), moq.It.isAny(), moq.It.isAny())).returns(() => queryReaderMock.object);
-    queryReaderMock.setup(async (x) => x.step()).returns(async () => false);
+    imodelMock.setup((x) => x.categories).returns(() => categoriesMock.object);
+    queryReaderMock.setup(async (x) => x.toArray()).returns(async () => [{ id: categoryId }]);
+    categoriesMock.setup(async (x) => x.getCategoryInfo([categoryId])).returns(async () => categoriesInfo);
     viewManagerMock.setup((x) => x.selectedView).returns(() => selectedViewMock.object);
     selectedViewMock.setup((x) => x.view).returns(() => selectedViewStateMock.object);
     selectedViewMock.setup((x) => x.perModelCategoryVisibility).returns(() => perModelCategoryVisibilityMock.object);
-    viewportMock.setup((x) => x.view).returns(() => viewStateMock.object);
-    viewStateMock.setup((x) => x.is3d()).returns(() => true);
     perModelCategoryVisibilityMock.setup((x) => x[Symbol.iterator]()).returns(() => [][Symbol.iterator]());
+    viewportMock.setup((x) => x.view).returns(() => viewStateMock.object);
+    // TODO: remove this eslint rule when tree-widget uses itwinjs-core 4.0.0 version
+    viewStateMock.setup((x) => x.is3d()).returns(() => true); // eslint-disable-line @itwin/no-internal
+  });
+
+  afterEach(() => {
+    perModelCategoryVisibilityMock.reset();
+    viewManagerMock.reset();
+    imodelMock.reset();
+    queryReaderMock.reset();
+    categoriesMock.reset();
+    viewportMock.reset();
+    viewStateMock.reset();
+    selectedViewMock.reset();
+    selectedViewStateMock.reset();
+    sinon.restore();
   });
 
   describe("toggleAllCategories", () => {
-
-    beforeEach(() => {
-      imodelMock.reset();
-      queryReaderMock.reset();
-      queryRowProxyMock.reset();
-
-      imodelMock.setup((x) => x.createQueryReader(moq.It.isAny(), moq.It.isAny(), moq.It.isAny())).returns(() => queryReaderMock.object);
-      queryReaderMock.setup(async (x) => x.step()).returns(async () => true);
-      queryReaderMock.setup(async (x) => x.step()).returns(async () => false);
-      queryReaderMock.setup((x) => x.current).returns(() => queryRowProxyMock.object);
-      queryRowProxyMock.setup((x) => x.id).returns(() => "CategoryId");
-      imodelMock.setup((x) => x.subcategories).returns(() => subCategoriesCacheMock.object);
-    });
-
     it("enables all categories", async () => {
       await toggleAllCategories(viewManagerMock.object, imodelMock.object, true, viewportMock.object, false);
       selectedViewMock.verify((x) => x.changeCategoryDisplay(["CategoryId"], true, moq.It.isAny()), moq.Times.once());
@@ -117,101 +104,84 @@ describe("CategoryVisibilityUtils", () => {
   });
 
   describe("enableCategory", () => {
-
-    beforeEach(() => {
-      perModelCategoryVisibilityMock.reset();
-      selectedViewMock.reset();
-      imodelMock.reset();
-      subCategoriesCacheMock.reset();
-
-      selectedViewMock.setup((x) => x.view).returns(() => selectedViewStateMock.object);
-      selectedViewMock.setup((x) => x.perModelCategoryVisibility).returns(() => perModelCategoryVisibilityMock.object);
-      imodelMock.setup((x) => x.subcategories).returns(() => subCategoriesCacheMock.object);
-      subCategoriesCacheMock.setup((x) => x.getSubCategories("CategoryId")).returns(() => new Set(categories[0].subCategoryIds));
-      perModelCategoryVisibilityMock.setup((x) => x[Symbol.iterator]()).returns(() => [][Symbol.iterator]());
-    });
-
-    it("enables category", () => {
-      enableCategory(viewManagerMock.object, imodelMock.object, ["CategoryId"], true, false, false);
+    it("enables category", async () => {
+      await enableCategory(viewManagerMock.object, imodelMock.object, ["CategoryId"], true, false, false);
       selectedViewMock.verify((x) => x.changeCategoryDisplay(["CategoryId"], true, false), moq.Times.once());
     });
 
-    it("disables category", () => {
-      enableCategory(viewManagerMock.object, imodelMock.object, ["CategoryId"], false, false, false);
+    it("disables category", async () => {
+      await enableCategory(viewManagerMock.object, imodelMock.object, ["CategoryId"], false, false, false);
       selectedViewMock.verify((x) => x.changeCategoryDisplay(["CategoryId"], false, false), moq.Times.once());
     });
 
-    it("disables category and subcategories", () => {
-      enableCategory(viewManagerMock.object, imodelMock.object, ["CategoryId"], false, false, true);
+    it("disables category and subcategories", async () => {
+      await enableCategory(viewManagerMock.object, imodelMock.object, ["CategoryId"], false, false, true);
       selectedViewMock.verify((x) => x.changeCategoryDisplay(["CategoryId"], false, true), moq.Times.once());
-      selectedViewMock.verify((x) => x.changeSubCategoryDisplay("SubCategoryId", false), moq.Times.once());
+      selectedViewMock.verify((x) => x.changeSubCategoryDisplay(moq.It.isAny(), moq.It.isAny()), moq.Times.once());
     });
 
-    it("removes overrides per model when enabling category", () => {
+    it("removes overrides per model when enabling category", async () => {
       const ovrs = [{ modelId: "ModelId", categoryId: "CategoryId", visible: false }];
       perModelCategoryVisibilityMock.reset();
       perModelCategoryVisibilityMock.setup((x) => x[Symbol.iterator]()).returns(() => ovrs[Symbol.iterator]());
-      enableCategory(viewManagerMock.object, imodelMock.object, ["CategoryId"], true, false, false);
+      await enableCategory(viewManagerMock.object, imodelMock.object, ["CategoryId"], true, false, false);
       selectedViewMock.verify((x) => x.changeCategoryDisplay(["CategoryId"], true, false), moq.Times.once());
       perModelCategoryVisibilityMock.verify((x) => x.setOverride(["ModelId"], ["CategoryId"], PerModelCategoryVisibility.Override.None), moq.Times.once());
     });
 
-    it("does not change category state if selectedView is undefined", () => {
+    it("does not change category state if selectedView is undefined", async () => {
       viewManagerMock.reset();
       viewManagerMock.setup((x) => x.selectedView).returns(() => undefined);
-      enableCategory(viewManagerMock.object, imodelMock.object, ["CategoryId"], false, false, false);
+      await enableCategory(viewManagerMock.object, imodelMock.object, ["CategoryId"], false, false, false);
       selectedViewMock.verify((x) => x.changeCategoryDisplay(["CategoryId"], false, false), moq.Times.never());
     });
 
-    it("enables category in all viewports", () => {
+    it("enables category in all viewports", async () => {
       viewStateMock.reset();
-      viewStateMock.setup((x) => x.is3d()).returns(() => true);
+      // TODO: remove this eslint rule when tree-widget uses itwinjs-core 4.0.0 version
+      viewStateMock.setup((x) => x.is3d()).returns(() => true); // eslint-disable-line @itwin/no-internal
       const otherViewMock = moq.Mock.ofType<Viewport>();
       otherViewMock.setup((x) => x.view).returns(() => viewStateMock.object);
       otherViewMock.setup((x) => x.perModelCategoryVisibility).returns(() => perModelCategoryVisibilityMock.object);
       mockViewManagerForEachViewport(otherViewMock.object);
       selectedViewStateMock.setup((x) => x.is3d()).returns(() => true);
 
-      enableCategory(viewManagerMock.object, imodelMock.object, ["CategoryId"], true, true, false);
+      await enableCategory(viewManagerMock.object, imodelMock.object, ["CategoryId"], true, true, false);
       viewManagerMock.verifyAll();
       otherViewMock.verify((x) => x.changeCategoryDisplay(["CategoryId"], true, false), moq.Times.once());
     });
 
-    it("disables category in all viewports", () => {
+    it("disables category in all viewports", async () => {
       viewStateMock.reset();
-      viewStateMock.setup((x) => x.is3d()).returns(() => true);
+      // TODO: remove this eslint rule when tree-widget uses itwinjs-core 4.0.0 version
+      viewStateMock.setup((x) => x.is3d()).returns(() => true); // eslint-disable-line @itwin/no-internal
       const otherViewMock = moq.Mock.ofType<Viewport>();
       otherViewMock.setup((x) => x.view).returns(() => viewStateMock.object);
       otherViewMock.setup((x) => x.perModelCategoryVisibility).returns(() => perModelCategoryVisibilityMock.object);
       mockViewManagerForEachViewport(otherViewMock.object, moq.Times.exactly(2));
       selectedViewStateMock.setup((x) => x.is3d()).returns(() => true);
 
-      enableCategory(viewManagerMock.object, imodelMock.object, ["CategoryId"], false, true, false);
+      await enableCategory(viewManagerMock.object, imodelMock.object, ["CategoryId"], false, true, false);
       viewManagerMock.verifyAll();
       otherViewMock.verify((x) => x.changeCategoryDisplay(["CategoryId"], false, false), moq.Times.once());
     });
 
-    it("does not change category if viewport and selected view has different types", () => {
+    it("does not change category if viewport and selected view has different types", async () => {
       viewStateMock.reset();
-      viewStateMock.setup((x) => x.is3d()).returns(() => false);
+      // TODO: remove this eslint rule when tree-widget uses itwinjs-core 4.0.0 version
+      viewStateMock.setup((x) => x.is3d()).returns(() => false); // eslint-disable-line @itwin/no-internal
       const otherViewMock = moq.Mock.ofType<Viewport>();
       otherViewMock.setup((x) => x.view).returns(() => viewStateMock.object);
       otherViewMock.setup((x) => x.perModelCategoryVisibility).returns(() => perModelCategoryVisibilityMock.object);
       mockViewManagerForEachViewport(otherViewMock.object);
       selectedViewStateMock.setup((x) => x.is3d()).returns(() => true);
 
-      enableCategory(viewManagerMock.object, imodelMock.object, ["CategoryId"], false, true, false);
+      await enableCategory(viewManagerMock.object, imodelMock.object, ["CategoryId"], false, true, false);
       otherViewMock.verify((x) => x.changeCategoryDisplay(["CategoryId"], false, false), moq.Times.never());
     });
   });
 
   describe("enableSubCategory", () => {
-
-    beforeEach(() => {
-      selectedViewMock.reset();
-      selectedViewMock.setup((x) => x.view).returns(() => selectedViewStateMock.object);
-    });
-
     it("enables subCategory", () => {
       enableSubCategory(viewManagerMock.object, "SubCategoryId", true);
       selectedViewMock.verify((x) => x.changeSubCategoryDisplay("SubCategoryId", true), moq.Times.once());
@@ -232,7 +202,8 @@ describe("CategoryVisibilityUtils", () => {
 
     it("enables subCategory in all viewports", () => {
       viewStateMock.reset();
-      viewStateMock.setup((x) => x.is3d()).returns(() => true);
+      // TODO: remove this eslint rule when tree-widget uses itwinjs-core 4.0.0 version
+      viewStateMock.setup((x) => x.is3d()).returns(() => true); // eslint-disable-line @itwin/no-internal
       const otherViewMock = moq.Mock.ofType<Viewport>();
       otherViewMock.setup((x) => x.view).returns(() => viewStateMock.object);
       mockViewManagerForEachViewport(otherViewMock.object);
@@ -245,7 +216,8 @@ describe("CategoryVisibilityUtils", () => {
 
     it("disables subCategory in all viewports", () => {
       viewStateMock.reset();
-      viewStateMock.setup((x) => x.is3d()).returns(() => true);
+      // TODO: remove this eslint rule when tree-widget uses itwinjs-core 4.0.0 version
+      viewStateMock.setup((x) => x.is3d()).returns(() => true); // eslint-disable-line @itwin/no-internal
       const otherViewMock = moq.Mock.ofType<Viewport>();
       otherViewMock.setup((x) => x.view).returns(() => viewStateMock.object);
       mockViewManagerForEachViewport(otherViewMock.object);
@@ -258,7 +230,8 @@ describe("CategoryVisibilityUtils", () => {
 
     it("does not change subCategory state if viewport and selectedView has different types", () => {
       viewStateMock.reset();
-      viewStateMock.setup((x) => x.is3d()).returns(() => false);
+      // TODO: remove this eslint rule when tree-widget uses itwinjs-core 4.0.0 version
+      viewStateMock.setup((x) => x.is3d()).returns(() => false); // eslint-disable-line @itwin/no-internal
       const otherViewMock = moq.Mock.ofType<Viewport>();
       otherViewMock.setup((x) => x.view).returns(() => viewStateMock.object);
       mockViewManagerForEachViewport(otherViewMock.object);
@@ -268,6 +241,5 @@ describe("CategoryVisibilityUtils", () => {
       viewManagerMock.verifyAll();
       otherViewMock.verify((x) => x.changeSubCategoryDisplay("SubCategoryId", false), moq.Times.never());
     });
-
   });
 });