+++ packages/itwin/tree-widget/CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log - @itwin/tree-widget-react
 
-This log was last generated on Fri, 24 Mar 2023 10:46:34 GMT and should not be manually modified.
+This log was last generated on Mon, 03 Apr 2023 15:34:07 GMT and should not be manually modified.
+
+## 0.8.0
+Mon, 03 Apr 2023 15:34:07 GMT
+
+### Minor changes
+
+- Removed `SpatialTree` and related components in favor of the ones in `@itwin/breakdown-trees-react`
+- Updated dependencies (`itwinui@2`, `appui@4`, `presentation-components@4`, `itwinjs-core@3.7`)
+
+### Patches
+
+- Use iTwin UI components instead of custom ones and CSS variables from `@itwin/itwinui-variables` instead of `@itwin/core-react`
 
 ## 0.7.2
 Fri, 24 Mar 2023 10:46:34 GMT
@@ -201,6 +213,10 @@ Tue, 13 Jul 2021 17:43:28 GMT
 ### Minor changes
 
 - Added a UiProvider that implements the existing tree widget
+
+### Updates
+
+- add rebuild script
 
 ## 1.3.1
 Tue, 08 Jun 2021 21:23:59 GMT